# openapi.yaml
...
openapi: 3.0.3
info:
  title: Scorvia
  version: 0.1.0
  description: >-
    Governed AutoML credit-decision factory with interpretability, regime, imbalance, and production promotion gates.
  x-product-slug: ai-fin-scorvia
  x-source: ai-in-financial/aiinfinance-190206170721/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/candidates/{candidateId}/evaluations:
    parameters:
      - $ref: "#/components/parameters/CandidateId"
    post:
      operationId: runEvaluation
      tags: [Evaluations]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeEvaluationReport"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CandidateId:
      name: candidateId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
...
    EvaluationReport:
      type: object
      required: [id, candidateId, passed]
      properties:
        id: { type: string }
        candidateId: { type: string }
        passed: { type: boolean }
        metrics: { type: object, additionalProperties: true }
        regimeTestsPassed: { type: boolean }
        imbalanceAware: { type: boolean }
    EvaluationCreate:
      type: object
      properties:
        includeRegimeTests: { type: boolean, default: true }
        requireImbalanceMetrics: { type: boolean, default: true }
...
    DataEnvelopeEvaluationReport:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/EvaluationReport"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...