# platform/api-server/src/domains/evaluations/handlers/evaluation.handlers.ts
/**
 * Handlers for Evaluation
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: evaluations
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { EvaluationsDomainModule } from "../dependencies/evaluations-ddd.dependencies.js";
import type { GetEvaluationInput, ListEvaluationsInput, RunEvaluationInput } from "@scorvia/services/evaluations";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/candidates/{candidateId}/evaluations
 * Operation: runEvaluation
 */
export async function runEvaluation(
  request: FastifyRequest<{ Params: { candidateId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: EvaluationsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RunEvaluationInput = {
    candidateId: request.params.candidateId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.evaluations.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/candidates/{candidateId}/evaluations
 * Operation: listEvaluations
 */
export async function listEvaluations(
  request: FastifyRequest<{ Params: { candidateId: string }; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: EvaluationsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListEvaluationsInput = {
    candidateId: request.params.candidateId,
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.evaluations.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for GET /v1/evaluations/{evaluationId}
 * Operation: getEvaluation
 */
export async function getEvaluation(
  request: FastifyRequest<{ Params: { evaluationId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: EvaluationsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetEvaluationInput = {
    evaluationId: request.params.evaluationId,
  };
  const result = await deps.useCases.evaluations.get.execute(input);
  return reply.code(200).send(result);
}